py/pyproject.toml
[project]
name = "py"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.2",
    "resonate-sdk>=0.6.4",
]
